requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==1.0.0
astroid==3.3.5
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
blinker==1.8.2
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.9
distlib==0.3.9
docutils==0.21.2
filelock==3.16.1
Flask==3.0.3
frozenlist==1.4.1
identify==2.6.1
idna==3.10
imagesize==1.4.1
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
multidict==6.1.0
nodeenv==1.9.1
packaging==24.1
platformdirs==4.3.6
pre_commit==4.0.1
Pygments==2.18.0
PyJWT==2.9.0
pylint==3.3.1
PyMySQL==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
ruff==0.6.9
snowballstemmer==2.2.0
Sphinx==8.1.0
sphinx-rtd-theme==3.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
tomlkit==0.13.2
twilio==9.3.2
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.6
Werkzeug==3.0.4
yarl==1.13.1
